# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Oficina Mecânica API",
    "version": "1.0.0",
    "description": "MVP — Sistema Integrado de Atendimento"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "summary": "Login",
        "tags": [
          "Autenticação"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "senha": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "email",
                  "senha"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "usuario": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "nome": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "ADMIN",
                            "FUNCIONARIO"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "nome",
                        "email",
                        "role"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "token",
                    "usuario"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/auth/usuarios": {
      "post": {
        "summary": "Criar usuário (Admin)",
        "tags": [
          "Autenticação"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 200
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "senha": {
                    "type": "string",
                    "minLength": 8,
                    "allOf": [
                      {
                        "pattern": "[A-Z]"
                      },
                      {
                        "pattern": "[0-9]"
                      },
                      {
                        "pattern": "[@$!%*?&]"
                      }
                    ]
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "ADMIN",
                      "FUNCIONARIO"
                    ],
                    "default": "FUNCIONARIO"
                  }
                },
                "required": [
                  "nome",
                  "email",
                  "senha"
                ],
                "additionalProperties": false
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "nome": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "criadoEm": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "nome",
                    "email",
                    "role",
                    "criadoEm"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Listar usuários (Admin)",
        "tags": [
          "Autenticação"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string",
                        "enum": [
                          "ADMIN",
                          "FUNCIONARIO"
                        ]
                      },
                      "ativo": {
                        "type": "boolean"
                      },
                      "criadoEm": {
                        "type": "string",
                        "format": "date-time"
                      }
                    },
                    "required": [
                      "id",
                      "nome",
                      "email",
                      "role",
                      "ativo",
                      "criadoEm"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Dados do usuário logado",
        "tags": [
          "Autenticação"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sub": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "ADMIN",
                        "FUNCIONARIO",
                        "CLIENTE"
                      ]
                    },
                    "email": {
                      "type": "string"
                    },
                    "nome": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "sub",
                    "role"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}